moved files (#110)
* moved tools stuff from Animation and Utility to new Tools folder
* moved debug stuff into Utility
