+++ QSB/Events/PlayerJoin.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections;
 using QSB.Messaging;
 using QSB.TransformSync;
+using QSB.Utility;
 using UnityEngine;
 using UnityEngine.Networking;
