+++ QSB/Events/PlayerLeave.cs
@@ -1,5 +1,6 @@
 ﻿using System.Linq;
 using QSB.Messaging;
+using QSB.Utility;
 using UnityEngine;
 using UnityEngine.Networking;
