+++ QSB/PlayerInfo.cs
@@ -1,5 +1,5 @@
 ﻿using System.Linq;
-using QSB.Animation;
+using QSB.Tools;
 using QSB.Utility;
 using UnityEngine;
