+++ QSB/QSB.csproj
@@ -113,10 +113,10 @@
     <Compile Include="Animation\AnimFloatParam.cs" />
     <Compile Include="Animation\AnimTriggerMessage.cs" />
     <Compile Include="Animation\AnimTrigger.cs" />
-    <Compile Include="Animation\QSBFlashlight.cs" />
-    <Compile Include="Animation\QSBTool.cs" />
-    <Compile Include="Animation\ToolType.cs" />
-    <Compile Include="DebugActions.cs" />
+    <Compile Include="Tools\QSBFlashlight.cs" />
+    <Compile Include="Tools\QSBTool.cs" />
+    <Compile Include="Tools\ToolType.cs" />
+    <Compile Include="Utility\DebugActions.cs" />
     <Compile Include="Events\EventListener.cs" />
     <Compile Include="Events\EventType.cs" />
     <Compile Include="Events\FullStateMessage.cs" />
@@ -128,7 +128,7 @@
     <Compile Include="Events\JoinMessage.cs" />
     <Compile Include="Events\LeaveMessage.cs" />
     <Compile Include="Events\PlayerLeave.cs" />
-    <Compile Include="DebugLog.cs" />
+    <Compile Include="Utility\DebugLog.cs" />
     <Compile Include="Events\DeathMessage.cs" />
     <Compile Include="Events\PlayerJoin.cs" />
     <Compile Include="Events\PlayerMessage.cs" />
@@ -144,7 +144,7 @@
     <Compile Include="TransformSync\PlayerCameraSync.cs" />
     <Compile Include="Utility\FlagsHelper.cs" />
     <Compile Include="TransformSync\PlayerHUDMarker.cs" />
-    <Compile Include="Utility\PlayerToolsManager.cs" />
+    <Compile Include="Tools\PlayerToolsManager.cs" />
     <Compile Include="Utility\QuaternionHelper.cs" />
     <Compile Include="TimeSync\PreserveTimeScale.cs" />
     <Compile Include="TransformSync\ShipTransformSync.cs" />
