+++ QSB/TimeSync/RespawnOnDeath.cs
@@ -3,6 +3,7 @@ using OWML.ModHelper.Events;
 using QSB.Events;
 using QSB.Messaging;
 using QSB.TransformSync;
+using QSB.Utility;
 using UnityEngine;
 
 namespace QSB.TimeSync
