+++ QSB/TransformSync/SectorSync.cs
@@ -1,6 +1,7 @@
 ﻿using QSB.Messaging;
 using UnityEngine;
 using System.Linq;
+using QSB.Utility;
 using UnityEngine.SceneManagement;
 
 namespace QSB.TransformSync
