+++ QSB/Utility/DebugLog.cs
@@ -1,10 +1,10 @@
-﻿using OWML.Common;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
+using OWML.Common;
 using UnityEngine;
 using UnityEngine.UI;
 
-namespace QSB
+namespace QSB.Utility
 {
     public class DebugLog : MonoBehaviour
     {
